Refactor talismanrc and cmd
