+++ cmd/runner.go
@@ -49,7 +49,6 @@ func (r *Runner) RunWithoutErrors(promptContext prompt.PromptContext) int {
 
 //Scan scans git commit history for potential secrets and returns 0 or 1 as exit code
 func (r *Runner) Scan(reportDirectory string, tRC *talismanrc.TalismanRC, ignoreHistory bool) int {
-
 	fmt.Printf("\n\n")
 	utility.CreateArt("Running Scan..")
 	additions := scanner.GetAdditions(ignoreHistory)
